feat: (#149) 관리자 전용 학식 추가/수정 기능 구현
