+++ src/entities/lunch-menu/api/lunchMenuListQueries.ts
@@ -1,0 +1,9 @@
+import { queryOptions } from '@tanstack/react-query';
+import { getLunchMenus } from './lunchMenuList';
+import { lunchMenuQueryKeys } from './lunchMenuQueryKeys';
+
+export const lunchMenuListQueryOptions = () =>
+  queryOptions({
+    queryKey: lunchMenuQueryKeys.lists(),
+    queryFn: getLunchMenus,
+  });
