+++ src/entities/lunch-menu/api/lunchMenuQueryKeys.ts
@@ -1,0 +1,4 @@
+export const lunchMenuQueryKeys = {
+  all: () => ['lunchMenus'] as const,
+  lists: () => [...lunchMenuQueryKeys.all(), 'list'] as const,
+};
