+++ src/pages/main/ui/layout/main-tab-section/ui/MainTabSection.tsx
@@ -1,15 +1,14 @@
-import { useMemo, useState } from 'react';
-import { useQuery } from '@tanstack/react-query';
+import { useMemo } from 'react';
+import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { Plus } from 'lucide-react';
 import type { PostSyncRequest } from '@/entities/post';
 import { myUserQueryOptions } from '@/entities/user';
+import { lunchMenuListQueryOptions, lunchMenuQueryKeys, type MainLunchMenu } from '@/entities/lunch-menu';
 import type { MainTab } from '../../main-tabs/model/types';
 import LunchSection from '@/widgets/lunch-section';
 import PostSection from '@/widgets/post-section';
 import RankingSection from '@/widgets/ranking-section';
 import RoomSection from '@/widgets/room-section';
-import { mockLunchMenus } from '@/widgets/lunch-section/model/mockLunchMenus';
-import type { MainLunchMenu } from '@/widgets/lunch-section/model/types';
 import type { MainRankingItem } from '@/widgets/ranking-section/model/types';
 
 interface MainTabSectionProps {
@@ -43,11 +42,13 @@ const MainTabSection = ({
   onPostSyncHandled,
   onRequireLogin,
 }: MainTabSectionProps) => {
-  const [lunchMenus, setLunchMenus] = useState<MainLunchMenu[]>(mockLunchMenus);
+  const queryClient = useQueryClient();
   const isRoomTab = activeTab === 'ROOM';
   const isPostTab = activeTab === 'POST';
   const myUserQuery = useQuery(myUserQueryOptions());
   const isAdmin = myUserQuery.data?.role === 'ADMIN';
+  const lunchMenusQuery = useQuery(lunchMenuListQueryOptions());
+  const lunchMenus = useMemo(() => lunchMenusQuery.data ?? [], [lunchMenusQuery.data]);
 
   const rankings = useMemo<MainRankingItem[]>(
     () =>
@@ -65,8 +66,16 @@ const MainTabSection = ({
     [lunchMenus],
   );
 
+  const updateLunchMenusCache = (
+    updater: (menus: MainLunchMenu[]) => MainLunchMenu[],
+  ) => {
+    queryClient.setQueryData<MainLunchMenu[]>(lunchMenuQueryKeys.lists(), (currentMenus) =>
+      updater(currentMenus ?? []),
+    );
+  };
+
   const handleLike = (menuId: number) => {
-    setLunchMenus((currentMenus) =>
+    updateLunchMenusCache((currentMenus) =>
       currentMenus.map((menu) => {
         if (menu.id !== menuId) {
           return menu;
@@ -87,7 +96,7 @@ const MainTabSection = ({
   };
 
   const handleDislike = (menuId: number) => {
-    setLunchMenus((currentMenus) =>
+    updateLunchMenusCache((currentMenus) =>
       currentMenus.map((menu) => {
         if (menu.id !== menuId) {
           return menu;
@@ -105,18 +114,6 @@ const MainTabSection = ({
         };
       }),
     );
-  };
-
-  const handleLunchMenuSaved = (menu: MainLunchMenu) => {
-    setLunchMenus((currentMenus) => {
-      const existingIndex = currentMenus.findIndex((item) => item.id === menu.id);
-
-      if (existingIndex === -1) {
-        return [menu, ...currentMenus];
-      }
-
-      return currentMenus.map((item) => (item.id === menu.id ? menu : item));
-    });
   };
 
   return (
@@ -148,7 +145,6 @@ const MainTabSection = ({
           isAdmin={isAdmin}
           onLike={handleLike}
           onDislike={handleDislike}
-          onMenuSaved={handleLunchMenuSaved}
         />
       ) : null}
       {activeTab === 'RANKING' ? <RankingSection rankings={rankings} /> : null}
